Refactor: Add theme folder in design-system. Change references for COLOR_MAP.
