+++ packages/design-system/src/components/circle/stories/circle.stories.tsx
@@ -20,7 +20,7 @@ import type { Meta, StoryObj } from '@storybook/react';
 /**
  * Internal dependencies
  */
-import { COLOR_MAP } from '../../../constants';
+import { COLOR_MAP } from '../../../theme/colors';
 
 /**
  * Internal dependencies.
