+++ packages/design-system/src/index.ts
@@ -15,6 +15,7 @@
  */
 
 export * from './icons';
+export * from './theme/colors';
 export * from './components';
 export * from './utils';
 export * from './constants';
