+++ packages/design-system/src/theme/colors.ts
@@ -1,0 +1,48 @@
+/*
+ * Copyright 2023 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+export const COLOR_MAP: { [key: string]: string } = {
+  functional: '#5CC971',
+  marketing: '#F3AE4E',
+  analytics: '#4C79F4',
+  uncategorized: '#EC7159',
+  brightGray: '#E8EAED',
+  mediumGray: '#BDBDBD',
+  SecureOnly: '#A98307',
+  DomainMismatch: '#7D8471',
+  NotOnPath: '#79553D',
+  SameSiteStrict: '#AF2B1E',
+  SameSiteLax: '#FFA420',
+  SameSiteUnspecifiedTreatedAsLax: '#FF7514',
+  SameSiteNoneInsecure: '#924E7D',
+  UserPreferences: '#E63244',
+  ThirdPartyPhaseout: '#999950',
+  ThirdPartyBlockedInFirstPartySet: '#DE4C8A',
+  UnknownError: '#DE4C8A',
+  SchemefulSameSiteStrict: '#763C28',
+  SchemefulSameSiteLax: '#C6A664',
+  SchemefulSameSiteUnspecifiedTreatedAsLax: '#E7EBDA',
+  SamePartyFromCrossPartyContext: '#008F39',
+  NameValuePairExceedsMaxSize: '#1C542D',
+  ExcludeSameSiteUnspecifiedTreatedAsLax: '#1C542D',
+  ExcludeSameSiteNoneInsecure: '#EDFF21',
+  ExcludeSameSiteLax: '#A5A5A5',
+  ExcludeSameSiteStrict: '#256D7B',
+  ExcludeInvalidSameParty: '#898176',
+  ExcludeSamePartyCrossPartyContext: '#497E76',
+  ExcludeDomainNonASCII: '#7FB5B5',
+  ExcludeThirdPartyCookieBlockedInFirstPartySet: '#EFA94A',
+  ExcludeThirdPartyPhaseout: '#CF3476',
+};
