+++ packages/extension/src/view/popup/components/legend/stories/legend.stories.tsx
@@ -17,7 +17,7 @@
  * External dependencies.
  */
 import type { Meta, StoryObj } from '@storybook/react';
-import { COLOR_MAP } from '@ps-analysis-tool/common';
+import { COLOR_MAP } from '@ps-analysis-tool/design-system';
 
 /**
  * Internal dependencies.
